+++ packages/terra-avatar/src/terra-dev-site/doc/example/AvatarThreeInitials.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Avatar from '../../../Avatar';
+import Avatar from 'terra-avatar';
 
 const AvatarThreeInitials = () => (
   <Avatar ariaLabel="John Doe" initials="JSR" />