+++ packages/terra-form-checkbox/tests/nightwatch/HiddenLabelCheckbox.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import Checkbox from '../../lib/Checkbox';
-
-const checkbox = () => (<Checkbox id="hidden" labelText="can you see me?" isLabelHidden />);
-
-export default checkbox;